+++ fabric/src/main/java/io/github/sakurawald/fuji/core/manager/impl/scheduler/ScheduleManager.java
@@ -33,6 +33,7 @@ public class ScheduleManager extends BaseManager {
 
     public static final String CRON_EVERY_SECOND = "* * * ? * *";
     public static final String CRON_EVERY_MINUTE = "0 * * ? * * *";
+    public static final String CRON_EVERY_FIVE_SECONDS = "0/5 * * ? * * *";
 
     private static final Set<BaseJob> RESCHEDULABLE_JOBS = new HashSet<>();