+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/works/WorksInitializer.java
@@ -8,19 +8,16 @@ import io.github.sakurawald.fuji.core.config.handler.abst.BaseConfigurationHandl
 import io.github.sakurawald.fuji.core.config.handler.impl.ObjectConfigurationHandler;
 import io.github.sakurawald.fuji.core.event.impl.ServerLifecycleEvents;
 import io.github.sakurawald.fuji.core.manager.Managers;
-import io.github.sakurawald.fuji.core.manager.impl.scheduler.ScheduleManager;
 import io.github.sakurawald.fuji.module.initializer.ModuleInitializer;
 import io.github.sakurawald.fuji.module.initializer.works.config.model.WorksConfigModel;
 import io.github.sakurawald.fuji.module.initializer.works.config.model.WorksDataModel;
 import io.github.sakurawald.fuji.module.initializer.works.gui.WorksGui;
-import io.github.sakurawald.fuji.module.initializer.works.job.WorksScheduleJob;
+import io.github.sakurawald.fuji.module.initializer.works.job.WorksOnScheduleDispatcherJob;
 import io.github.sakurawald.fuji.module.initializer.works.structure.work.abst.Work;
-import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.network.ServerPlayerEntity;
-import org.quartz.JobDataMap;
 
 @Document("""
-    Provides a bill-board, for players to post and share their works.
+    Provides a `bill-board`, for `players` to post and share their works.
     """)
 public class WorksInitializer extends ModuleInitializer {
 
@@ -29,10 +26,11 @@ public class WorksInitializer extends ModuleInitializer {
 
     public static final BaseConfigurationHandler<WorksConfigModel> config = new ObjectConfigurationHandler<>(BaseConfigurationHandler.CONFIG_JSON, WorksConfigModel.class);
 
-    @Document("Open works GUI.")
+    @Document("Open the works GUI.")
     @CommandNode("works")
     private static int $works(@CommandSource ServerPlayerEntity player) {
-        new WorksGui(player, works.model().works, 0).open();
+        new WorksGui(player, works.model().works, 0)
+            .open();
         return CommandHelper.Return.SUCCESS;
     }
 
@@ -44,12 +42,8 @@ public class WorksInitializer extends ModuleInitializer {
     @Override
     protected void onInitialize() {
         ServerLifecycleEvents.SERVER_STARTED.register(server -> {
-            WorksScheduleJob worksScheduleJob = new WorksScheduleJob(new JobDataMap() {
-                {
-                    this.put(MinecraftServer.class.getName(), server);
-                }
-            }, () -> ScheduleManager.CRON_EVERY_MINUTE);
-            Managers.getScheduleManager().scheduleJob(worksScheduleJob);
+            WorksOnScheduleDispatcherJob job = WorksOnScheduleDispatcherJob.makeInstance();
+            Managers.getScheduleManager().scheduleJob(job);
         });
     }